Changed TStates from ints to bytes.

+++ TuringMachines/BinaryInvertTuringMachine.cs
@@ -1,17 +1,17 @@
 ﻿namespace TuringMachines;
 
-public sealed class BinaryInvertTuringMachine : TuringMachine<int, char>
+public sealed class BinaryInvertTuringMachine : TuringMachine<byte, char>
 {
     protected override IReadOnlySet<char> Alphabet { get; } = new HashSet<char> { '0', '1' };
 
-    protected override int InitialState { get; } = 1;
+    protected override byte InitialState { get; } = 1;
 
-    protected override IReadOnlyDictionary<int, bool> FinalStates { get; } = new Dictionary<int, bool>
+    protected override IReadOnlyDictionary<byte, bool> FinalStates { get; } = new Dictionary<byte, bool>
     {
         { 7, true },
     };
 
-    protected override IReadOnlyDictionary<(int, char), TMAction> TransitionFunctions { get; } = new Dictionary<(int, char), TMAction>
+    protected override IReadOnlyDictionary<(byte, char), TMAction> TransitionFunctions { get; } = new Dictionary<(byte, char), TMAction>
     {
         { (1, Blank), new MoveRight { ToState = 2 } },
